fix run_until_completion() return stack handling
Previously, there would be occasional but reproducible crashes for more
complex programs. It turns out that using the undocumented "XSprePUSH"
macro fixes this. Xsubpp uses the macro for CODE but not PPCODE
sections. Some distros on CPAN use that macro habitually. The XS docs
don't mention this macro. It seems that using this macro sometimes
prevents a kind of stack corruption. E.g. without it, `$_[0]` would
sometimes appear as the sole return value even when other values were
pushed.

The mysteries of XS.